Fix unpost bug, add multi-channel posting

+++ hubbot.js
@@ -634,9 +634,11 @@ bot.on("messageReactionRemove", async (msg, emoji, user) => {
 })
 
 bot.on("messageDelete", async (msg) => {
+	if(!msg.channel.guild) return;
 	try {
 		bot.db.query(`DELETE FROM reactposts WHERE server_id=? AND channel_id=? AND message_id=?`,[msg.channel.guild.id, msg.channel.id, msg.id]);
 		await bot.utils.deleteTicketPost(bot, msg.channel.guild.id, msg.channel.id, msg.id);
+		await bot.utils.deletePost(bot, msg.channel.guild.id, msg.id);
 
 		var log = await bot.utils.getBanLogByMessage(bot, msg.channel.guild.id, msg.channel.id, msg.id);
 		if(log) await bot.utils.removeBanLog(bot, log.hid, msg.channel.guild.id);
